[Task] - Refactoring the data generation archives to another directory
