+++ src/main/java/com/example/api/dealership/core/domain/AddressModel.java
@@ -1,8 +1,12 @@
 package com.example.api.dealership.core.domain;
 
 import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.hibernate.annotations.GenericGenerator;
 
 import jakarta.persistence.Entity;
@@ -11,7 +15,9 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
 @Entity
+@Builder
 @Data
+@AllArgsConstructor
 @NoArgsConstructor
 @Table(name="TB_ADDRESS")
 public class AddressModel {
